+++ test/phoenix/controller/render_test.exs
@@ -117,6 +117,30 @@ defmodule Phoenix.Controller.RenderTest do
     assert conn.resp_body == "Hello\n"
   end
 
+  test "render/3 renders with View and Template with atom for template" do
+    conn = put_in conn.params["format"], "json"
+    conn = put_in conn.private[:phoenix_action], :show
+    conn = put_view(conn, nil)
+    conn = render(conn, MyApp.UserView, :show)
+    assert conn.resp_body == ~s({"foo":"bar"})
+  end
+
+  test "render/3 renders with View and Template" do
+    conn = put_in conn.params["format"], "json"
+    conn = put_in conn.private[:phoenix_action], :show
+    conn = put_view(conn, nil)
+    conn = render(conn, MyApp.UserView, "show.json")
+    assert conn.resp_body == ~s({"foo":"bar"})
+  end
+
+  test "render/4 renders with View and Template" do
+    conn = put_in conn.params["format"], "html"
+    conn = put_in conn.private[:phoenix_action], :index
+    conn = put_view(conn, nil)
+    conn = render(conn, MyApp.UserView, "index.html", title: "Hello")
+    assert conn.resp_body == "Hello\n"
+  end
+
   test "errors when rendering without format" do
     assert_raise RuntimeError, ~r/cannot render template :index because conn.params/, fn ->
       render(conn(), :index)